This adds benchmarks for static, state parameter and network obtained data processing. In general, there is little low-hanging fruit except for the final JSON format clean-up done by eval, demonstrated by the benchmark comparisons below. This code is not run in filebeat, but does hint at optimisations that could be done there.

```
$ go test -run ^$ -bench . -count 10 > bench_not_fast.out
$ go test -run ^$ -bench . -count 10 -fast_marshal > bench_fast.out
$ benchstat bench_not_fast.out bench_fast.out
goos: darwin
goarch: amd64
pkg: github.com/elastic/mito
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
                                  │ bench_not_fast.out │           bench_fast.out            │
                                  │       sec/op       │   sec/op     vs base                │
Mito/hello_world_static-16                1466.5n ± 3%   651.6n ± 5%  -55.57% (p=0.000 n=10)
Mito/hello_world_object_static-16          5.408µ ± 4%   3.163µ ± 3%  -41.51% (p=0.000 n=10)
Mito/nested_static-16                      17.88µ ± 5%   12.22µ ± 2%  -31.67% (p=0.000 n=10)
Mito/encode_json_static-16                 22.43µ ± 2%   21.30µ ± 4%   -5.06% (p=0.002 n=10)
Mito/nested_collate_static-16              5.465µ ± 3%   4.405µ ± 3%  -19.39% (p=0.000 n=10)
Mito/hello_world_state-16                 1559.5n ± 3%   751.9n ± 2%  -51.79% (p=0.000 n=10)
Mito/hello_world_object_state-16           5.821µ ± 4%   3.355µ ± 2%  -42.37% (p=0.000 n=10)
Mito/nested_state-16                       16.30µ ± 3%   10.85µ ± 1%  -33.45% (p=0.000 n=10)
Mito/encode_json_state-16                  4.898µ ± 3%   3.501µ ± 2%  -28.53% (p=0.000 n=10)
Mito/nested_collate_list_state-16          14.61µ ± 3%   13.37µ ± 7%   -8.54% (p=0.002 n=10)
Mito/nested_collate_map_state-16           14.76µ ± 3%   13.49µ ± 3%   -8.59% (p=0.000 n=10)
Mito/null_net-16                           78.49µ ± 8%   72.04µ ± 3%   -8.22% (p=0.000 n=10)
Mito/hello_world_net-16                    82.69µ ± 3%   81.37µ ± 3%        ~ (p=0.315 n=10)
Mito/hello_world_object_net-16             95.39µ ± 1%   92.60µ ± 4%   -2.92% (p=0.007 n=10)
Mito/nested_net-16                         111.0µ ± 3%   102.7µ ± 2%   -7.45% (p=0.000 n=10)
Mito/encode_json_null_net-16               86.98µ ± 7%   79.97µ ± 3%   -8.06% (p=0.001 n=10)
Mito/encode_json_net-16                    94.39µ ± 4%   91.97µ ± 4%        ~ (p=0.280 n=10)
Mito/nested_collate_list_net-16            107.0µ ± 3%   106.8µ ± 5%        ~ (p=1.000 n=10)
Mito/nested_collate_map_net-16             107.4µ ± 3%   107.0µ ± 6%        ~ (p=1.000 n=10)
geomean                                    21.32µ        16.80µ       -21.19%
```

Please take a look.